ByteOrder.h: simplify compiler test
* A problem with our gcc requires adding casts for gcc4 when
  the __builtin_bswap functions are used with a format string
* Unlike gcc2, the __builtin_bswap functions do not get disabled
  despite using -fno-builtins, hence added compiler check in
  runtime_loader/utility.cpp

+++ src/add-ons/kernel/busses/scsi/ahci/ahci_port.cpp
@@ -819,8 +819,8 @@ AHCIPort::ScsiUnmap(scsi_ccb* request, scsi_unmap_parameter_list* unmapBlocks)
 dprintf("TRIM SCSI:\n");
 for (uint32 i = 0; i < scsiRangeCount; i++) {
 	dprintf("[%3" B_PRIu32 "] %" B_PRIu64 " : %" B_PRIu32 "\n", i,
-		B_BENDIAN_TO_HOST_INT64(unmapBlocks->blocks[i].lba),
-		B_BENDIAN_TO_HOST_INT32(unmapBlocks->blocks[i].block_count));
+		(uint64)B_BENDIAN_TO_HOST_INT64(unmapBlocks->blocks[i].lba),
+		(uint32)B_BENDIAN_TO_HOST_INT32(unmapBlocks->blocks[i].block_count));
 }
 
 	uint32 scsiIndex = 0;
